backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary

# Caching
redis==5.0.1
django-redis==5.4.0

# Task Queue
celery==5.3.4

# Storage
Pillow

# API Documentation
drf-yasg==1.21.7

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment Variables
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
